+++ Changelog.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## Added
+
+- Implemented `SizeOf` for `core::any::TypeId`
+
 ## Changed
 
 - Made `SizeOf` impls for `Context`, `TotalSize` and `HumanBytes` unconditional
   (were previously dependent on the `derive` feature)
-- Implemented `SizeOf` for `core::any::TypeId`
+- `derive(SizeOf)` now partially normalizes types and doesn't emit bounds for `fn` types
 
 ## [0.1.3] - 2022-10-21